libk tweaks

+++ libk/stdlib.h
@@ -1,5 +1,6 @@
 #ifndef STDLIB_H
 #define STDLIB_H
+#include <stddef.h>
 #include <stdint.h>
-void itoa(int64_t x, char *str);
+void itoa(size_t x, char *str);
 #endif
