+++ libk/string.h
@@ -2,8 +2,8 @@
 #define STRING_H
 #include <stdint.h>
 #include <stddef.h>
-uint32_t strlen(const char *str);
-uint32_t strnlen(const char *str, size_t n);
+size_t strlen(const char *str);
+size_t strnlen(const char *str, size_t n);
 char *strcat(char *dst, const char *src);
 char *strncat(char *dst, const char *src, size_t n);
 char *strcpy(char *dst, const char *src);
